move fec_timer.py
