+++ spinn_front_end_common/interface/provenance/__init__.py
@@ -17,6 +17,7 @@ from .abstract_provides_local_provenance_data import (
     AbstractProvidesLocalProvenanceData)
 from .abstract_provides_provenance_data_from_machine import (
     AbstractProvidesProvenanceDataFromMachine)
+from .fec_timer import FecTimer
 from .log_store_db import LogStoreDB
 from .provenance_reader import ProvenanceReader
 from .provides_provenance_data_from_machine_impl import (
